requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
openai==0.27.8
requests==2.31.0
websockets==11.0.3
python-multipart==0.0.6
SQLAlchemy==2.0.22
aiofiles==23.2.1 